refactor: split chat_handler.go and local.go by concern
chat_handler.go (1309 → 3 files):
- chat_handler.go: struct, WebSocket, connection handling, providers
- chat_handler_message.go: handleMessage LLM tool-calling loop
- chat_handler_tools.go: SQL and MCP tool execution

local.go (1305 → 4 files):
- local.go: core operations (upload, download, delete, list, copy, move)
- local_bucket.go: bucket CRUD
- local_signed.go: signed URL generation and validation
- local_chunked.go: chunked upload sessions
