+++ internal/ai/chat_handler.go
@@ -6,7 +6,6 @@ import (
 	"fmt"
 	"strings"
 	"sync"
-	"time"
 
 	"github.com/gofiber/contrib/v3/websocket"
 	"github.com/gofiber/fiber/v3"
@@ -390,717 +389,6 @@ func (h *ChatHandler) handleStartChat(ctx context.Context, chatCtx *ChatContext,
 		Str("conversation_id", state.ID).
 		Str("chatbot", chatbot.Name).
 		Msg("Chat session started")
-}
-
-// handleMessage handles a user message
-func (h *ChatHandler) handleMessage(ctx context.Context, chatCtx *ChatContext, msg *ClientMessage) {
-	start := time.Now()
-
-	state := chatCtx.Conversations[msg.ConversationID]
-	if state == nil {
-		h.sendError(chatCtx, msg.ConversationID, "NO_SESSION", "No active chat session")
-		return
-	}
-
-	chatbot := chatCtx.ActiveChatbot
-	if chatbot == nil {
-		h.sendError(chatCtx, msg.ConversationID, "NO_CHATBOT", "No active chatbot")
-		return
-	}
-
-	// Resolve template variables in chatbot annotation values (e.g., http-allowed-domains)
-	if err := h.ResolveChatbotTemplates(ctx, chatbot, chatCtx.UserID); err != nil {
-		log.Warn().Err(err).Str("chatbot", chatbot.Name).Msg("Failed to resolve chatbot templates")
-		// Continue with unresolved values - don't fail the request
-	}
-
-	// Determine user identifier for rate limiting
-	userIdentifier := "anonymous"
-	if chatCtx.UserID != nil {
-		userIdentifier = *chatCtx.UserID
-	}
-
-	// Check per-minute rate limit
-	if !h.limiter.CheckRateLimit(chatbot.ID, userIdentifier, chatbot.RateLimitPerMinute) {
-		h.sendError(chatCtx, msg.ConversationID, "RATE_LIMITED", "Rate limit exceeded. Please try again later.")
-		return
-	}
-
-	// Check daily request limit
-	if !h.limiter.CheckDailyRequestLimit(chatbot.ID, userIdentifier, chatbot.DailyRequestLimit) {
-		h.sendError(chatCtx, msg.ConversationID, "DAILY_LIMIT", "Daily request limit exceeded.")
-		return
-	}
-
-	// Check turn limit
-	if state.TurnCount >= chatbot.MaxConversationTurns {
-		h.sendError(chatCtx, msg.ConversationID, "TURN_LIMIT", "Conversation turn limit reached")
-		return
-	}
-
-	// Send thinking progress
-	h.sendProgress(chatCtx, msg.ConversationID, "thinking", "Thinking...")
-
-	// Build system prompt with schema
-	userID := ""
-	if chatCtx.UserID != nil {
-		userID = *chatCtx.UserID
-	}
-
-	systemPrompt, err := h.schemaBuilder.BuildSystemPrompt(ctx, chatbot, userID)
-	if err != nil {
-		log.Error().Err(err).Msg("Failed to build system prompt")
-		h.sendError(chatCtx, msg.ConversationID, "PROMPT_ERROR", "Failed to build prompt")
-		return
-	}
-
-	// Retrieve RAG context if available (with user isolation)
-	if h.ragService != nil {
-		ragOpts := RetrieveContextOptions{
-			ChatbotID: chatbot.ID,
-			Query:     msg.Content,
-			UserID:    userID,
-		}
-		if chatbot.RAGMaxChunks > 0 {
-			ragOpts.MaxChunks = chatbot.RAGMaxChunks
-		}
-		if chatbot.RAGSimilarityThreshold > 0 {
-			ragOpts.Threshold = chatbot.RAGSimilarityThreshold
-		}
-		ragSection, err := h.ragService.RetrieveContext(ctx, ragOpts)
-		if err != nil {
-			log.Warn().Err(err).Str("chatbot_id", chatbot.ID).Msg("Failed to retrieve RAG context")
-			// Continue without RAG - don't fail the request
-		} else if ragSection != nil && ragSection.FormattedContext != "" {
-			systemPrompt = systemPrompt + "\n\n" + ragSection.FormattedContext
-			log.Debug().
-				Str("chatbot_id", chatbot.ID).
-				Str("conversation_id", msg.ConversationID).
-				Int("rag_section_len", len(ragSection.FormattedContext)).
-				Msg("RAG context added to system prompt")
-		}
-	}
-
-	// Build messages for LLM
-	messages := []Message{
-		{Role: RoleSystem, Content: systemPrompt},
-	}
-
-	// Add conversation history
-	messages = append(messages, state.Messages...)
-
-	// Add user message
-	userMsg := Message{Role: RoleUser, Content: msg.Content}
-	messages = append(messages, userMsg)
-
-	// Get provider
-	provider, err := h.getProvider(ctx, chatbot)
-	if err != nil {
-		log.Error().Err(err).Msg("Failed to get provider")
-		h.sendError(chatCtx, msg.ConversationID, "PROVIDER_ERROR", "AI provider not available")
-		return
-	}
-
-	// Save user message to conversation
-	_ = h.conversations.AddMessage(ctx, msg.ConversationID, userMsg, 0, 0)
-
-	// Tool calling loop - continue until AI generates content without tool calls
-	var totalUsage UsageStats
-	var accumulatedQueryResults []QueryResult // Accumulate query results for persistence
-	maxIterations := chatbot.MaxToolIterations
-	if maxIterations <= 0 {
-		maxIterations = 5
-	}
-
-	// Track consecutive tool validation failures to detect stubborn LLM behavior
-	var lastFailedTool string
-	var consecutiveFailures int
-	const maxConsecutiveFailures = 2
-
-	// Track whether think tool has been called (for enforcing ReAct pattern)
-	hasUsedThink := false
-	thinkRequired := chatbot.ReasoningMode == "react" || chatbot.ReasoningMode == "strict"
-
-	for iteration := 0; iteration < maxIterations; iteration++ {
-		// Determine forbidden tools based on user message and intent rules
-		var forbiddenTools []string
-		if len(chatbot.IntentRules) > 0 {
-			intentValidator := NewIntentValidator(chatbot.IntentRules, chatbot.RequiredColumns, chatbot.DefaultTable)
-			forbiddenTools = intentValidator.GetForbiddenTools(msg.Content)
-			if len(forbiddenTools) > 0 {
-				log.Debug().
-					Strs("forbidden_tools", forbiddenTools).
-					Str("user_message", msg.Content).
-					Msg("Filtering out forbidden tools based on intent rules")
-			}
-		}
-
-		// Helper to check if a tool is forbidden
-		isToolForbidden := func(toolName string) bool {
-			for _, ft := range forbiddenTools {
-				if ft == toolName {
-					return true
-				}
-			}
-			return false
-		}
-
-		// Helper to check if think tool is available
-		hasThinkTool := func(tools []Tool) bool {
-			for _, t := range tools {
-				if t.Function.Name == "think" {
-					return true
-				}
-			}
-			return false
-		}
-
-		// Build tools list based on chatbot configuration
-		var tools []Tool
-
-		// Add MCP tools if configured (includes execute_sql as MCP tool now)
-		if chatbot.HasMCPTools() && h.mcpExecutor != nil {
-			mcpToolDefs := h.mcpExecutor.GetAvailableTools(chatbot)
-			for _, def := range mcpToolDefs {
-				// Skip forbidden tools
-				if isToolForbidden(def.Name) {
-					continue
-				}
-				tools = append(tools, Tool{
-					Type:     "function",
-					Function: ToolFunction(def),
-				})
-			}
-		} else if !isToolForbidden("execute_sql") {
-			// Fallback: add legacy execute_sql if no MCP tools configured
-			tools = append(tools, ExecuteSQLTool)
-		}
-
-		// Enforce ReAct pattern: require think tool before other tools
-		// If reasoning mode is react/strict and think hasn't been used yet,
-		// only allow the think tool on the first iteration
-		if thinkRequired && !hasUsedThink && iteration == 0 && hasThinkTool(tools) {
-			// Filter to only include think tool
-			var thinkOnlyTools []Tool
-			for _, t := range tools {
-				if t.Function.Name == "think" {
-					thinkOnlyTools = append(thinkOnlyTools, t)
-					break
-				}
-			}
-			if len(thinkOnlyTools) > 0 {
-				tools = thinkOnlyTools
-				log.Debug().
-					Str("chatbot", chatbot.Name).
-					Msg("ReAct mode: restricting to think tool only on first iteration")
-			}
-		}
-
-		log.Debug().
-			Str("chatbot", chatbot.Name).
-			Int("total_tools", len(tools)).
-			Bool("think_required", thinkRequired).
-			Bool("has_used_think", hasUsedThink).
-			Msg("Tools available for chatbot")
-
-		// Create chat request
-		chatReq := &ChatRequest{
-			Messages:    messages,
-			MaxTokens:   chatbot.MaxTokens,
-			Temperature: chatbot.Temperature,
-			Tools:       tools,
-			Stream:      true,
-		}
-
-		// Track response for this iteration
-		var responseContent strings.Builder
-		var pendingToolCalls []ToolCall
-
-		// Stream callback
-		callback := func(event StreamEvent) error {
-			switch event.Type {
-			case "content":
-				responseContent.WriteString(event.Delta)
-				h.send(chatCtx, ServerMessage{
-					Type:           "content",
-					ConversationID: msg.ConversationID,
-					Delta:          event.Delta,
-				})
-
-			case "tool_call":
-				// Collect tool calls to execute after streaming completes
-				if event.ToolCall != nil {
-					toolName := event.ToolCall.FunctionName
-					// Accept legacy tools, MCP tools, or any tool requested by the model
-					pendingToolCalls = append(pendingToolCalls, ToolCall{
-						ID:   event.ToolCall.ID,
-						Type: "function",
-						Function: FunctionCall{
-							Name:      toolName,
-							Arguments: event.ToolCall.ArgumentsDelta,
-						},
-					})
-				}
-
-			case "done":
-				if event.Usage != nil {
-					totalUsage.PromptTokens += event.Usage.PromptTokens
-					totalUsage.CompletionTokens += event.Usage.CompletionTokens
-					totalUsage.TotalTokens += event.Usage.TotalTokens
-				}
-			}
-			return nil
-		}
-
-		// Stream the response
-		h.sendProgress(chatCtx, msg.ConversationID, "generating", "Generating response...")
-
-		if err := provider.ChatStream(ctx, chatReq, callback); err != nil {
-			log.Error().Err(err).Msg("Chat stream error")
-			h.sendError(chatCtx, msg.ConversationID, "STREAM_ERROR", "Error generating response")
-
-			if h.metrics != nil {
-				h.metrics.RecordAIChatRequest(chatbot.Name, "error", time.Since(start))
-			}
-			return
-		}
-
-		// If no tool calls, we're done
-		if len(pendingToolCalls) == 0 {
-			// Save assistant message with accumulated query results
-			assistantMsg := Message{
-				Role:         RoleAssistant,
-				Content:      responseContent.String(),
-				QueryResults: accumulatedQueryResults,
-			}
-			_ = h.conversations.AddMessage(ctx, msg.ConversationID, assistantMsg, totalUsage.PromptTokens, totalUsage.CompletionTokens)
-			break
-		}
-
-		// Add assistant message with tool calls to conversation
-		assistantMsg := Message{
-			Role:      RoleAssistant,
-			Content:   responseContent.String(),
-			ToolCalls: pendingToolCalls,
-		}
-		messages = append(messages, assistantMsg)
-
-		// Execute each tool call and add results
-		for _, tc := range pendingToolCalls {
-			toolName := tc.Function.Name
-
-			// Track if think tool was used (for ReAct pattern)
-			if toolName == "think" {
-				hasUsedThink = true
-			}
-
-			// Validate tool call against intent rules (requiredTool/forbiddenTool)
-			if len(chatbot.IntentRules) > 0 {
-				intentValidator := NewIntentValidator(chatbot.IntentRules, chatbot.RequiredColumns, chatbot.DefaultTable)
-				toolValidation := intentValidator.ValidateToolCall(msg.Content, toolName)
-
-				log.Debug().
-					Int("intent_rules_count", len(chatbot.IntentRules)).
-					Str("tool", toolName).
-					Str("user_message", msg.Content).
-					Bool("valid", toolValidation.Valid).
-					Strs("matched_keywords", toolValidation.MatchedKeywords).
-					Msg("Tool validation check")
-
-				if !toolValidation.Valid {
-					// Track consecutive failures for the same tool
-					if toolName == lastFailedTool {
-						consecutiveFailures++
-					} else {
-						lastFailedTool = toolName
-						consecutiveFailures = 1
-					}
-
-					// If the same tool fails too many times, break the loop
-					if consecutiveFailures >= maxConsecutiveFailures {
-						log.Warn().
-							Str("tool", toolName).
-							Int("failures", consecutiveFailures).
-							Msg("Breaking loop due to repeated tool validation failures")
-
-						h.send(chatCtx, ServerMessage{
-							Type:  "error",
-							Error: "Unable to process this request - the AI kept trying to use a tool that isn't allowed for this type of query. Please rephrase your question.",
-						})
-						return
-					}
-
-					// Build list of alternative tools (exclude the forbidden one)
-					var alternativeTools []string
-					for _, t := range chatbot.MCPTools {
-						if t != toolName {
-							alternativeTools = append(alternativeTools, t)
-						}
-					}
-
-					errMsg := fmt.Sprintf("TOOL NOT ALLOWED: %s. %s Available tools: %s. Please use one of these tools instead.",
-						strings.Join(toolValidation.Errors, "; "),
-						strings.Join(toolValidation.Suggestions, " "),
-						strings.Join(alternativeTools, ", "))
-
-					log.Debug().
-						Strs("errors", toolValidation.Errors).
-						Strs("alternative_tools", alternativeTools).
-						Str("error_message", errMsg).
-						Msg("Tool validation failed, returning error to LLM")
-					toolMsg := Message{
-						Role:       RoleTool,
-						Content:    errMsg,
-						ToolCallID: tc.ID,
-						Name:       toolName,
-					}
-					messages = append(messages, toolMsg)
-					continue // Skip execution, let AI retry with correct tool
-				}
-			}
-
-			toolResult, queryResult := h.executeToolCall(ctx, chatCtx, msg.ConversationID, chatbot, &tc, userID, msg.Content)
-
-			// Accumulate successful query results for persistence
-			if queryResult != nil {
-				accumulatedQueryResults = append(accumulatedQueryResults, *queryResult)
-			}
-
-			// Add tool result message
-			toolMsg := Message{
-				Role:       RoleTool,
-				Content:    toolResult,
-				ToolCallID: tc.ID,
-				Name:       tc.Function.Name,
-			}
-			messages = append(messages, toolMsg)
-		}
-
-		// Continue loop to get AI's response to tool results
-		log.Debug().
-			Int("iteration", iteration+1).
-			Int("tool_calls", len(pendingToolCalls)).
-			Msg("Processed tool calls, continuing conversation")
-	}
-
-	// Track token usage for daily budget enforcement
-	if chatbot.DailyTokenBudget > 0 {
-		userIdentifier := "anonymous"
-		if chatCtx.UserID != nil {
-			userIdentifier = *chatCtx.UserID
-		}
-		h.limiter.AddTokenUsage(chatbot.ID, userIdentifier, totalUsage.TotalTokens)
-	}
-
-	// Send completion
-	h.send(chatCtx, ServerMessage{
-		Type:           "done",
-		ConversationID: msg.ConversationID,
-		Usage:          &totalUsage,
-	})
-
-	// Record metrics
-	if h.metrics != nil {
-		h.metrics.RecordAIChatRequest(chatbot.Name, "success", time.Since(start))
-		h.metrics.RecordAITokens(chatbot.Name, totalUsage.PromptTokens, totalUsage.CompletionTokens)
-	}
-
-	log.Debug().
-		Str("conversation_id", msg.ConversationID).
-		Int("prompt_tokens", totalUsage.PromptTokens).
-		Int("completion_tokens", totalUsage.CompletionTokens).
-		Msg("Message processed")
-}
-
-// / executeToolCall executes a tool call and returns:
-// - the result as a string for the AI
-// - the QueryResult for persistence (nil if query failed or not a SQL query)
-func (h *ChatHandler) executeToolCall(ctx context.Context, chatCtx *ChatContext, conversationID string, chatbot *Chatbot, toolCall *ToolCall, userID string, userMessage string) (string, *QueryResult) {
-	toolName := toolCall.Function.Name
-
-	// Check if this is an MCP tool
-	if chatbot.HasMCPTools() && h.mcpExecutor != nil && IsToolAllowed(toolName, chatbot.MCPTools) {
-		return h.executeMCPTool(ctx, chatCtx, conversationID, chatbot, toolCall)
-	}
-
-	// Dispatch based on tool name for legacy tools
-	switch toolName {
-	case "execute_sql":
-		return h.executeSQLTool(ctx, chatCtx, conversationID, chatbot, toolCall, userID, userMessage)
-	default:
-		return fmt.Sprintf("Error: Unknown tool '%s'", toolName), nil
-	}
-}
-
-// executeSQLTool handles the execute_sql tool call
-func (h *ChatHandler) executeSQLTool(ctx context.Context, chatCtx *ChatContext, conversationID string, chatbot *Chatbot, toolCall *ToolCall, userID string, userMessage string) (string, *QueryResult) {
-	// Parse arguments
-	var args struct {
-		SQL         string `json:"sql"`
-		Description string `json:"description"`
-	}
-
-	if err := json.Unmarshal([]byte(toolCall.Function.Arguments), &args); err != nil {
-		log.Error().Err(err).Str("args", toolCall.Function.Arguments).Msg("Failed to parse SQL tool call arguments")
-		return fmt.Sprintf("Error: Failed to parse tool arguments: %v", err), nil
-	}
-
-	// Intent validation (before execution)
-	if len(chatbot.IntentRules) > 0 || len(chatbot.RequiredColumns) > 0 {
-		intentValidator := NewIntentValidator(
-			chatbot.IntentRules,
-			chatbot.RequiredColumns,
-			chatbot.DefaultTable,
-		)
-
-		// Pre-validate SQL to get tables accessed
-		preValidator := NewSQLValidator(chatbot.AllowedSchemas, chatbot.AllowedTables, chatbot.AllowedOperations)
-		preResult := preValidator.Validate(args.SQL)
-
-		// Validate intent matches query
-		intentResult := intentValidator.ValidateIntent(userMessage, args.SQL, preResult.TablesAccessed)
-		if !intentResult.Valid {
-			errMsg := fmt.Sprintf("Intent validation failed: %s", strings.Join(intentResult.Errors, "; "))
-			if len(intentResult.Suggestions) > 0 {
-				errMsg += fmt.Sprintf(" Suggestions: %s", strings.Join(intentResult.Suggestions, "; "))
-			}
-			log.Warn().
-				Str("user_message", userMessage).
-				Str("sql", args.SQL).
-				Strs("errors", intentResult.Errors).
-				Msg("Intent validation failed")
-			return errMsg, nil
-		}
-
-		// Validate required columns
-		colResult := intentValidator.ValidateRequiredColumns(args.SQL, preResult.TablesAccessed)
-		if !colResult.Valid {
-			errMsg := fmt.Sprintf("Required columns missing: %s", strings.Join(colResult.Errors, "; "))
-			if len(colResult.Suggestions) > 0 {
-				errMsg += fmt.Sprintf(" Suggestions: %s", strings.Join(colResult.Suggestions, "; "))
-			}
-			log.Warn().
-				Str("sql", args.SQL).
-				Strs("errors", colResult.Errors).
-				Msg("Required columns validation failed")
-			return errMsg, nil
-		}
-	}
-
-	// Send progress
-	h.sendProgress(chatCtx, conversationID, "querying", fmt.Sprintf("Executing: %s", args.Description))
-
-	// Execute SQL
-	execReq := &ExecuteRequest{
-		ChatbotName:       chatbot.Name,
-		ChatbotID:         chatbot.ID,
-		ConversationID:    conversationID,
-		UserID:            userID,
-		Role:              chatCtx.Role,
-		Claims:            chatCtx.Claims,
-		SQL:               args.SQL,
-		Description:       args.Description,
-		AllowedSchemas:    chatbot.AllowedSchemas,
-		AllowedTables:     chatbot.AllowedTables,
-		AllowedOperations: chatbot.AllowedOperations,
-	}
-
-	result, err := h.executor.Execute(ctx, execReq)
-	if err != nil {
-		log.Error().Err(err).Msg("SQL execution error")
-		return fmt.Sprintf("Error executing query: %v", err), nil
-	}
-
-	// Log to audit (unless execution logs are disabled)
-	if !chatbot.DisableExecutionLogs {
-		_ = h.auditLogger.LogFromExecuteResult(
-			ctx,
-			chatbot.ID, conversationID, "", userID,
-			args.SQL, result,
-			chatCtx.Role, chatCtx.IPAddress, chatCtx.UserAgent,
-		)
-
-		// Log to central logging service
-		if h.loggingService != nil {
-			h.loggingService.LogAI(ctx, map[string]any{
-				"tool":            "execute_sql",
-				"chatbot_id":      chatbot.ID,
-				"conversation_id": conversationID,
-				"success":         result.Success,
-				"rows_returned":   result.RowCount,
-				"tables":          result.TablesAccessed,
-				"duration_ms":     result.DurationMs,
-			}, "", userID)
-		}
-	}
-
-	// Send result to client for display
-	h.send(chatCtx, ServerMessage{
-		Type:           "query_result",
-		ConversationID: conversationID,
-		Query:          args.SQL,
-		Summary:        result.Summary,
-		RowCount:       result.RowCount,
-		Data:           result.Rows,
-	})
-
-	// Return result as string for AI to interpret
-	if !result.Success {
-		return fmt.Sprintf("Query failed: %s", result.Error), nil
-	}
-
-	// Build QueryResult for persistence
-	queryResult := &QueryResult{
-		Query:    args.SQL,
-		Summary:  result.Summary,
-		RowCount: result.RowCount,
-		Data:     result.Rows,
-	}
-
-	// Format result for AI - include summary and sample data
-	resultStr := fmt.Sprintf("Query executed successfully. %s\n", result.Summary)
-	if len(result.Rows) > 0 {
-		// Include first few rows as JSON for context
-		maxRows := 5
-		if len(result.Rows) < maxRows {
-			maxRows = len(result.Rows)
-		}
-		sampleData, _ := json.Marshal(result.Rows[:maxRows])
-		resultStr += fmt.Sprintf("Sample data (first %d rows): %s", maxRows, string(sampleData))
-	}
-
-	return resultStr, queryResult
-}
-
-// executeMCPTool handles MCP tool execution for chatbots with MCP tools configured
-func (h *ChatHandler) executeMCPTool(ctx context.Context, chatCtx *ChatContext, conversationID string, chatbot *Chatbot, toolCall *ToolCall) (string, *QueryResult) {
-	toolName := toolCall.Function.Name
-
-	// Parse tool arguments
-	var args map[string]any
-	if err := json.Unmarshal([]byte(toolCall.Function.Arguments), &args); err != nil {
-		log.Error().Err(err).Str("tool", toolName).Str("args", toolCall.Function.Arguments).Msg("Failed to parse MCP tool arguments")
-		return fmt.Sprintf("Error: Failed to parse tool arguments: %v", err), nil
-	}
-
-	// Send progress to client
-	progressMsg := fmt.Sprintf("Executing %s...", toolName)
-	if tableName, ok := args["table"].(string); ok && tableName != "" {
-		progressMsg = fmt.Sprintf("Executing %s on %s...", toolName, tableName)
-	}
-	h.sendProgress(chatCtx, conversationID, "executing", progressMsg)
-
-	// Execute the MCP tool
-	result, err := h.mcpExecutor.ExecuteTool(ctx, toolName, args, chatCtx, chatbot)
-	if err != nil {
-		log.Error().Err(err).Str("tool", toolName).Msg("MCP tool execution error")
-		return fmt.Sprintf("Error executing %s: %v", toolName, err), nil
-	}
-
-	if result.IsError {
-		log.Warn().Str("tool", toolName).Str("error", result.Content).Msg("MCP tool returned error")
-		return fmt.Sprintf("Error: %s", result.Content), nil
-	}
-
-	// Log successful execution
-	log.Debug().
-		Str("chatbot", chatbot.Name).
-		Str("tool", toolName).
-		Int("result_length", len(result.Content)).
-		Msg("MCP tool executed successfully")
-
-	// Parse query results for data-returning tools
-	var queryResult *QueryResult
-	switch toolName {
-	case "query_table":
-		queryResult = h.parseMCPQueryResult(toolName, args, result.Content)
-	case "execute_sql":
-		queryResult = h.parseMCPExecuteSQLResult(args, result.Content)
-	}
-
-	// Build server message
-	serverMsg := ServerMessage{
-		Type:           "tool_result",
-		ConversationID: conversationID,
-		Message:        toolName,
-	}
-
-	// Add structured fields for execute_sql
-	if toolName == "execute_sql" && queryResult != nil {
-		serverMsg.Query = queryResult.Query
-		serverMsg.Summary = queryResult.Summary
-		serverMsg.RowCount = queryResult.RowCount
-		serverMsg.Data = queryResult.Data
-	} else {
-		serverMsg.Data = []map[string]any{{"tool": toolName, "result": result.Content}}
-	}
-
-	// Suppress think/reasoning tool results from client when ShowReasoning is false
-	if toolName == "think" && !chatbot.ShowReasoning {
-		return result.Content, queryResult
-	}
-
-	h.send(chatCtx, serverMsg)
-
-	return result.Content, queryResult
-}
-
-// parseMCPQueryResult attempts to parse MCP query results for persistence
-func (h *ChatHandler) parseMCPQueryResult(toolName string, args map[string]any, resultContent string) *QueryResult {
-	// Try to parse the result as JSON array
-	var rows []map[string]any
-	if err := json.Unmarshal([]byte(resultContent), &rows); err != nil {
-		// Not valid JSON array, skip persistence
-		return nil
-	}
-
-	// Build a description for the query
-	tableName := ""
-	if t, ok := args["table"].(string); ok {
-		tableName = t
-	}
-
-	return &QueryResult{
-		Query:    fmt.Sprintf("MCP %s on %s", toolName, tableName),
-		Summary:  fmt.Sprintf("Query returned %d row(s)", len(rows)),
-		RowCount: len(rows),
-		Data:     rows,
-	}
-}
-
-// parseMCPExecuteSQLResult parses execute_sql MCP tool results for persistence
-func (h *ChatHandler) parseMCPExecuteSQLResult(args map[string]any, resultContent string) *QueryResult {
-	// Parse the result JSON from the MCP tool
-	var execResult struct {
-		Success    bool             `json:"success"`
-		RowCount   int              `json:"row_count"`
-		Columns    []string         `json:"columns"`
-		Rows       []map[string]any `json:"rows"`
-		Summary    string           `json:"summary"`
-		DurationMs int64            `json:"duration_ms"`
-		Tables     []string         `json:"tables"`
-	}
-
-	if err := json.Unmarshal([]byte(resultContent), &execResult); err != nil {
-		return nil
-	}
-
-	if !execResult.Success {
-		return nil
-	}
-
-	// Extract SQL query from tool arguments
-	sqlQuery := ""
-	if sql, ok := args["sql"].(string); ok {
-		sqlQuery = sql
-	}
-
-	return &QueryResult{
-		Query:    sqlQuery,
-		Summary:  execResult.Summary,
-		RowCount: execResult.RowCount,
-		Data:     execResult.Rows,
-	}
 }
 
 // handleCancel handles cancellation of a generation
